Refactoring displays to be accurate.
